Spell Queues
Move to spell queues for autocasting.
Creates generalized functions library to aid in code cleanup.

Queues will hold exactly one spell of each type, and the starting types
are run in this order:
 - priority (will be cast first, for stuff like cure on blindness/gouge)
 - heal (for healing spells, cast next)
 - any (for refresh, etc, to be cast at next opportunity)
 - combat (violence spells, only do it if we're not healing)
 - nco (non combat operations, shapechange, vigor, etc)

To add spells, use spellAdd {queue name} {complete casting command}, and for
the sake of consistency use the short syntax but full command, like
"shapechange frost".

Handles being bashed, on knees, and not awake.  There are undoubtedly
more casts to add.

Autoheal runs off group information, and casting heal or refresh updates
MSDP_GROUP to keep the information current and double casting to a
minimum.  Needs protection for self in the event that there is no group
at all.